usb: typec: tcpm: modify the conditions for port as sink
For a standard Type-A to Type-C cable, only one CC pin
has Rp, and the other CC pin is open, which is the condition
that the tcpm framework considers the port as sink mode.
However, for Huawei Type-A to Type-C cable, if both CC pins
have Rp, the tcpm framework considers the port as an unknown
mode, so maybe this situation is included in the port as a
condition for sink mode.

Change-Id: If03e8ec1298779ff3767b410238fe2b83431b0b4
Signed-off-by: Wang Jie <dave.wang@rock-chips.com>